+++ js/admin.elements.js
@@ -19,7 +19,6 @@ var fts_tesseract_elements = {
 	tesseract_threads: null,
 	tesseract_pdf: null,
 	tesseract_pdf_limit: null,
-	tesseract_pdf_skip_text: null,
 
 	init: function () {
 		fts_tesseract_elements.tesseract_div = $('#files_ocr-tesseract');
@@ -31,7 +30,6 @@ var fts_tesseract_elements = {
 		fts_tesseract_elements.tesseract_ocr = $('#tesseract_ocr');
 		fts_tesseract_elements.tesseract_pdf = $('#tesseract_pdf');
 		fts_tesseract_elements.tesseract_pdf_limit = $('#tesseract_pdf_limit');
-		fts_tesseract_elements.tesseract_pdf_skip_text = $('#tesseract_pdf_skip_text');
 
 		fts_tesseract_elements.tesseract_ocr.on('change', fts_tesseract_elements.updateSettings);
 		fts_tesseract_elements.tesseract_psm.on('change', fts_tesseract_elements.updateSettings);
@@ -45,9 +43,6 @@ var fts_tesseract_elements = {
 		fts_tesseract_elements.tesseract_threads.on('change', fts_tesseract_elements.updateSettings);
 		fts_tesseract_elements.tesseract_pdf.on('change', fts_tesseract_elements.updateSettings);
 		fts_tesseract_elements.tesseract_pdf_limit.on('change', fts_tesseract_elements.updateSettings);
-		fts_tesseract_elements.tesseract_pdf_skip_text.on(
-			'change', fts_tesseract_elements.updateSettings
-		);
 	},